requirements.txt
pyusb==1.1.1
pynput==1.7.7
pycryptodome==3.10.1
